исправил по замечаниям после проверки
