--- js/consts.js
@@ -1,13 +1,0 @@
-const TOTAL_PHOTO_AMOUNT = 25;
-
-const LIKES_AMOUNT = {
-  min: 15,
-  max: 200,
-};
-
-const COMMENTS_AMOUNT = {
-  min: 0,
-  max: 30,
-};
-
-export {TOTAL_PHOTO_AMOUNT, LIKES_AMOUNT, COMMENTS_AMOUNT};
