+++ js/mocks/mock-consts.js
@@ -1,0 +1,13 @@
+const TOTAL_PHOTO_AMOUNT = 25;
+
+const LikesAmount = {
+  MIN: 15,
+  MAX: 200,
+};
+
+const CommentsAmount = {
+  MIN: 0,
+  MAX: 30,
+};
+
+export {TOTAL_PHOTO_AMOUNT, LikesAmount, CommentsAmount};
